Only owner of a list can delete items

+++ lists/views.py
@@ -43,6 +43,8 @@ def add_item(request, list_id):
 def delete_item(request, item_id):
 	list_item = Item.objects.get(id=item_id)
 	list_ = list_item.list
+	if isinstance(request.user, AnonymousUser) and request.user is not list_.owner:
+			return redirect('/404_page_does_not_exist/')
 	list_item.delete()
 	return redirect('/lists/%d/' % (list_.id,))
